+++ integration/js/utils/WebdriverBrowserStack.js
@@ -1,6 +1,7 @@
 const {Builder} = require('selenium-webdriver');
 const {getBuildId, getRunDetails} = require('./BrowserStackLogs');
 const {AppPage} = require('../pages/AppPage');
+const {MeetingReadinessCheckerPage} = require('../pages/MeetingReadinessCheckerPage');
 
 const getOS = capabilities => {
   switch (capabilities.platform) {
@@ -78,7 +79,7 @@ const getBrowserStackUrl = () => {
 };
 
 class BrowserStackSession {
-  static async createSession(capabilities) {
+  static async createSession(capabilities, appName) {
     let cap = {};
     if (capabilities.browserName === 'chrome') {
       cap = getChromeCapabilities(capabilities);
@@ -89,11 +90,12 @@ class BrowserStackSession {
       .usingServer(getBrowserStackUrl())
       .withCapabilities(cap)
       .build();
-    return new BrowserStackSession(driver);
+    return new BrowserStackSession(driver, appName);
   }
 
-  constructor(inDriver) {
+  constructor(inDriver, appName) {
     this.driver = inDriver;
+    this.appName = appName;
   }
 
   async init() {
@@ -120,7 +122,9 @@ class BrowserStackSession {
 
   getAppPage() {
     if (this.page === undefined) {
-      this.page = new AppPage(this.driver, this.logger);
+      this.page = this.appName === 'meeting'
+        ? new AppPage(this.driver, this.logger)
+        : new MeetingReadinessCheckerPage(this.driver, this.logger);
     }
     return this.page;
   }