+++ src/index.ts
@@ -18,6 +18,15 @@ import BaseConnectionHealthPolicy from './connectionhealthpolicy/BaseConnectionH
 import BaseTask from './task/BaseTask';
 import BitrateParameters from './videouplinkbandwidthpolicy/BitrateParameters';
 import BrowserBehavior from './browserbehavior/BrowserBehavior';
+import CheckAudioConnectivityFeedback from './meetingreadinesschecker/CheckAudioConnectivityFeedback';
+import CheckAudioInputFeedback from './meetingreadinesschecker/CheckAudioInputFeedback';
+import CheckAudioOutputFeedback from './meetingreadinesschecker/CheckAudioOutputFeedback';
+import CheckCameraResolutionFeedback from './meetingreadinesschecker/CheckCameraResolutionFeedback';
+import CheckContentShareConnectivityFeedback from './meetingreadinesschecker/CheckContentShareConnectivityFeedback';
+import CheckNetworkTCPConnectivityFeedback from './meetingreadinesschecker/CheckNetworkTCPConnectivityFeedback';
+import CheckNetworkUDPConnectivityFeedback from './meetingreadinesschecker/CheckNetworkUDPConnectivityFeedback';
+import CheckVideoConnectivityFeedback from './meetingreadinesschecker/CheckVideoConnectivityFeedback';
+import CheckVideoInputFeedback from './meetingreadinesschecker/CheckVideoInputFeedback';
 import CleanRestartedSessionTask from './task/CleanRestartedSessionTask';
 import CleanStoppedSessionTask from './task/CleanStoppedSessionTask';
 import ClientMetricReport from './clientmetricreport/ClientMetricReport';
@@ -56,6 +65,7 @@ import DefaultJPEGDecoderComponentFactory from './jpegdecoder/DefaultJPEGDecoder
 import DefaultJPEGDecoderController from './jpegdecoder/controller/DefaultJPEGDecoderController';
 import DefaultJPEGDecoderInstance from './jpegdecoder/instance/DefaultJPEGDecoderInstance';
 import DefaultMediaDeviceFactory from './mediadevicefactory/DefaultMediaDeviceFactory';
+import DefaultMeetingReadinessChecker from './meetingreadinesschecker/DefaultMeetingReadinessChecker';
 import DefaultMeetingSession from './meetingsession/DefaultMeetingSession';
 import DefaultModality from './modality/DefaultModality';
 import DefaultPingPong from './pingpong/DefaultPingPong';
@@ -135,6 +145,7 @@ import MediaRecordingEvent from './mediarecording/MediaRecordingEvent';
 import MediaRecordingFactory from './mediarecording/MediaRecordingFactory';
 import MediaRecordingOptions from './mediarecording/MediaRecordingOptions';
 import MediaStreamBroker from './mediastreambroker/MediaStreamBroker';
+import MeetingReadinessChecker from './meetingreadinesschecker/MeetingReadinessChecker';
 import MeetingSession from './meetingsession/MeetingSession';
 import MeetingSessionConfiguration from './meetingsession/MeetingSessionConfiguration';
 import MeetingSessionCredentials from './meetingsession/MeetingSessionCredentials';
@@ -318,6 +329,15 @@ export {
   BaseTask,
   BitrateParameters,
   BrowserBehavior,
+  CheckAudioConnectivityFeedback,
+  CheckAudioInputFeedback,
+  CheckAudioOutputFeedback,
+  CheckCameraResolutionFeedback,
+  CheckContentShareConnectivityFeedback,
+  CheckNetworkTCPConnectivityFeedback,
+  CheckNetworkUDPConnectivityFeedback,
+  CheckVideoConnectivityFeedback,
+  CheckVideoInputFeedback,
   CleanRestartedSessionTask,
   CleanStoppedSessionTask,
   ClientMetricReport,
@@ -356,6 +376,7 @@ export {
   DefaultJPEGDecoderController,
   DefaultJPEGDecoderInstance,
   DefaultMediaDeviceFactory,
+  DefaultMeetingReadinessChecker,
   DefaultMeetingSession,
   DefaultModality,
   DefaultPingPong,
@@ -435,6 +456,7 @@ export {
   MediaRecordingFactory,
   MediaRecordingOptions,
   MediaStreamBroker,
+  MeetingReadinessChecker,
   MeetingSession,
   MeetingSessionConfiguration,
   MeetingSessionCredentials,