+++ src/meetingreadinesschecker/CheckContentShareConnectivityFeedback.ts
@@ -1,0 +1,12 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export enum CheckContentShareConnectivityFeedback {
+  Succeeded,
+  Failed,
+  PermissionDenied,
+  TimedOut,
+  ConnectionFailed,
+}
+
+export default CheckContentShareConnectivityFeedback;