+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add npm login and logout as part of publish script
+- Add Meeting Readiness Checker APIs
 
 ### Changed
 - Change WebRTC semantics to Unified Plan by default for Chromium-based browsers